requirements.txt
applicationinsights==0.11.7
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-tinymce4-lite==1.7.5
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.7.3.2
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
djangorestframework==3.11.0
requests==2.23.0
urllib3==1.25.8
django-nested-admin==3.3.0


